modularize the test data generating python script
