+++ .gitignore
@@ -9,3 +9,5 @@ node_modules/
 
 # build
 /lib/
+
+__pycache__/
